requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
google-cloud-storage==3.10.1
apache-airflow-providers-google
python-dotenv